requirements.txt
Flask
Flask-SQLAlchemy
flask-restful
Flask-Security
Flask-Migrate
email_validator